Moved MemoryRouterProps to common types
